Boxplots for Scotland and Pakistan cases
New notebooks for creating the boxplots for the Scotland and Pakistan cases. The netcdf files are readed one by one, because when computing the regional sources with the function calc_regional_sources it takes too much time if we work with a single dataset.

The results of WRF-WVTs are also included here, they are read as a csv and then added to the xarray dataset srcs_regional_combined.