Update combine_data.py to read WRF-WVT results from netCDF
Now WRF-WVT results can are readed from netCDF files. This allows the same functions to be used in Functions.py as with the other methods, but requires the netCDF files which are in the "results WRF-WVT" folders since the beginning of this month (before, there were only CSV files). Observe that you may want to omit WRF-WVT results when representing gridded results.