Update Functions.py to correct the calculation of regional_sources
At the last meeting a problem was mentioned with the calculation of regional sources using the calc_regional_sources function. I think we were weighting the area of gridded evaporative sources twice, since in the calc_regional_sources function we were using as weights ‘np.cos(np.deg2rad(sources[lat_name]))’ to take into account the decrease in area with increasing latitude, but at the same time we are passing as argument to this function the already area-weighted results, srcs_frac. Replacing the weights in calc_regional_sources by 1 can solve this issue and still provide the regional sources as fractions, as in BeeswarmAndBoxPlots_AllCases.ipynb